FEAT: imu basic implementation

+++ components/hw_drive/private_headers/mpu6050.h
@@ -7,9 +7,9 @@
 #include <stdint.h>
 
 typedef struct {
-	int16_t X;
-	int16_t Y;
-	int16_t Z;
+	float X;
+	float Y;
+	float Z;
 } S_Gyro;
 
 #define REGISTER_GYRO_X 0x43
